DT-2897 Allow unprotected queries so the UI introspection works. All real data queries still protected by role on controller
